Function amazon_sp_product_type_definitions::apis::definitions_api::get_definitions_product_type

pub async fn get_definitions_product_type(
    configuration: &Configuration,
    product_type: &str,
    marketplace_ids: Vec<String>,
    seller_id: Option<&str>,
    product_type_version: Option<&str>,
    requirements: Option<&str>,
    requirements_enforced: Option<&str>,
    locale: Option<&str>
) -> Result<ProductTypeDefinition, Error<GetDefinitionsProductTypeError>>

Retrieve an Amazon product type definition. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 5 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see Usage Plans and Rate Limits in the Selling Partner API.